+++ example/lib/detail/detail_page_handler_exception.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+sealed class DetailPageHandlerException {
+  static onError(
+    Object error,
+    StackTrace stackTrace,
+    BuildContext? context,
+    RouteSettings? routeSettings,
+  ) {
+    if (context != null) {
+      showDialog(
+        context: context,
+        builder: (context) => AlertDialog(
+          title: const Text('Error'),
+          content: Text(error.toString()),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(),
+              child: const Text('Ok'),
+            )
+          ],
+        ),
+      );
+    }
+  }
+}