+++ example/lib/home_async_value/model/user_model.dart
@@ -1,4 +1,0 @@
-class User {
-  String name;
-  User({required this.name});
-}