+++ example/lib/login/login_state.dart
@@ -1,7 +1,0 @@
-enum LoginStateStatus {
-  idle,
-  loading,
-  loaded,
-  needActivation,
-  error,
-}