+++ example/lib/main.dart
@@ -1,10 +1,8 @@
 import 'package:asyncstate/asyncstate.dart';
 import 'package:example/components/global_loading.dart';
 import 'package:example/detail/detail_page.dart';
-import 'package:example/home_async_value/home_async_value_page.dart';
-import 'package:example/login/login_page.dart';
+import 'package:example/detail/detail_page_handler_exception.dart';
 import 'package:example/profile/profile_page.dart';
-import 'package:example/root/root_page.dart';
 import 'package:flutter/material.dart';
 
 import 'home_loader/home_loader_page.dart';
@@ -22,51 +20,38 @@ class MyApp extends StatelessWidget {
     return AsyncStateBuilder(
       /// Here you can customize your default loading that will show every transaction
       /// Leave it and it will show a simple CircularProgress.adaptive indicator
-      loader: GlobalLoading(),
-      onError: (ErrorKit errorKit) {
+      loader: const GlobalLoading(),
+      onError: (error, stackTrace, context, routeSettings) {
         /// Here you can handle your exceptions
-        /// Leave it and it will show a simple SnackBar with the exception.toString()
-        ///
-        switch (errorKit.currentRoute) {
-          case '/Root/HomeLoader/Detail':
-            showDialog(
-              context: errorKit.context,
-              builder: (context) => AlertDialog(
-                title: const Text('Error'),
-                content: Text(errorKit.error.toString()),
-                actions: [
-                  TextButton(
-                    onPressed: () => Navigator.of(context).pop(),
-                    child: const Text('Ok'),
-                  )
-                ],
-              ),
+        switch (routeSettings?.name ?? '') {
+          case '/Detail':
+            DetailPageHandlerException.onError(
+              error,
+              stackTrace,
+              context,
+              routeSettings,
             );
             break;
           case _:
-            ScaffoldMessenger.of(errorKit.context).showSnackBar(
+            ScaffoldMessenger.of(context!).showSnackBar(
               SnackBar(
                 content: Text(
-                  errorKit.error.toString(),
+                  error.toString(),
                 ),
               ),
             );
             break;
         }
       },
-
       builder: (navigatorObserver) => MaterialApp(
         navigatorObservers: [navigatorObserver],
         themeMode: ThemeMode.dark,
         theme: ThemeData.dark(),
-        initialRoute: '/Root',
+        initialRoute: '/',
         routes: {
-          '/Root': (context) => const RootPage(),
-          '/Root/Login': (context) => LoginPage(),
-          '/Root/HomeAsyncValue': (context) => const HomeAsyncValuePage(),
-          '/Root/HomeLoader': (context) => const HomeLoaderPage(),
-          '/Root/HomeLoader/Detail': (context) => const DetailPage(),
-          '/Root/HomeLoader/ProfilePage': (context) => const ProfilePage(),
+          '/': (context) => const HomeLoaderPage(),
+          '/Detail': (context) => const DetailPage(),
+          '/Profile': (context) => const ProfilePage()
         },
       ),
     );