+++ lib/class/async_state_exception_handler.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-abstract interface class AsyncStateExceptionHandler {
-  void onException(
-    Object exception,
-    StackTrace stackTrace,
-    BuildContext context,
-  );
-}