+++ lib/enum/enum_loader_type.dart
@@ -1,1 +1,0 @@
-enum LoaderType { alertDialog, snackBar, materialBanner }